chore: Move PaymentBottomSheet to common component
